pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aios-insurance"
version = "0.1.0"
description = "Insurance Claims Processing - Vertical on AIOS Platform"
requires-python = ">=3.12"
dependencies = [
    "aios-core",
    "aios-ontology",
    "aios-data",
    "aios-deploy",
    "aios-api",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5.0,<6",
    "ruff>=0.4,<1",
    "pyright>=1.1,<2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/insurance"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
